Add police name data as JSON
Add the data to the repo, as a JSON array of names, sorted
alphabetically.

The process to retrieve new data is more manual than I would have liked,
but each step is documented below, along with why the long road was
taken.

Add script to fetch police data as html.

The seethroughny.net API only returns HTML, and the script must be run
in browser, but it should reliably get the data for further
transformation.

Add script to transform said html to json. This also runs in the
browser, on the page downloaded above.

This could all be potentially re-factored to a more automated,
server-oriented process, but this gets us the data for now.

Update `package.json` name.